+++ clientevents.cpp
@@ -172,11 +172,11 @@ void sendWantBuildList ( cBuilding *Building, cList<sBuildStruct*> *BuildList, b
 {
 	cNetMessage* message = new cNetMessage( GAME_EV_WANT_BUILDLIST );
 	message->pushBool ( bRepeat );
-	for (int i = (BuildList->Size()-1); i >= 0; i--)
+	for (int i = (int)BuildList->Size()-1; i >= 0; i--)
 	{
 		message->pushInt16((*BuildList)[i]->id);
 	}
-	message->pushInt16(BuildList->Size());
+	message->pushInt16((int)BuildList->Size());
 	message->pushInt16 ( Building->BuildSpeed );
 	message->pushInt16 ( Building->iID );
 	Client->sendNetMessage( message );
